:articulated_lorry: Renaming of files
Moved test utilities into their own directory
Renamed app utilities to lib to better fit naming convention
