+++ app/lib/gpDetailsParser.js
