+++ app/lib/gpOpeningTimesParser.js
