+++ test/unit/lib/getSampleResponse.js
